feat(metrics): we are recording timestamps to follow the whole translation of OpenAPI specs to TD
We are recording four timestamps: 1)received data from a standard Arrowhead pooling; 2) received a
OpenAPI spec from a service; 3) translated a OpenAPI spec to TD; 4) deployed the Web Thing. This
metrics are recorded in a file in the root of the project, and there is a line for each Web Thing
instanciated, the first value is the translated Arrowhead service ID. This way, we can record
metrics to a performance analysis experiment.